Separate store module. Update tsdoc comments
